Refactoring multipoint functions into their own module
