+++ src/alpss/multipoint/alpss_multipoint.py
@@ -1,8 +1,8 @@
 import logging
 from alpss.alpss_main import alpss_main
-from alpss.carrier.freq_refinement import find_carrier
-from alpss.io.reading import extract_data
-from alpss.io.multipoint_saving import save_combined_series
+from alpss.multipoint.freq_refinement import find_carrier
+from alpss.multipoint.extract_multipoint_data import extract_multipoint_data
+from alpss.multipoint.multipoint_saving import save_combined_series
 
 logger = logging.getLogger("alpss")
